refactor: 必要ないファイル削除＆viewディレクトリ作成_#76
