+++ components/view/commentComponent.tsx
@@ -1,12 +1,12 @@
 import React, { useContext } from "react";
 import "firebase/firestore";
-import { AppContext } from "../context/PageStates";
-import styles from "../components/scss/post.module.scss";
+import { AppContext } from "../../context/PageStates";
+import styles from "../../components/scss/post.module.scss";
 import { Button, TextField } from "@material-ui/core";
 import Stack from "@mui/material/Stack";
 import { Avatar } from "@mui/material";
 import { makeStyles } from "@material-ui/core/styles";
-import { useCommentAction } from "../hooks/useCommentAction";
+import { useCommentAction } from "../../hooks/useCommentAction";
 
 const useStyles = makeStyles((theme) => ({
   root: {
